Updated how the theme colours are managed
